Added docstrings LoginActionTest and UserRegisterTest. I refactored the UserRegisterTest because UserRegister itself is a UseCase, but also interacts with the classes in controller folders
