+++ src/test/java/UseCasesTest/UserRegister/UserRegisterTest.java
@@ -1,4 +1,4 @@
-package ControllerTest;
+package UseCasesTest.UserRegister;
 
 import DataBase.*;
 import UseCases.LoginAction.LoginAction;
@@ -12,7 +12,9 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 class UserRegisterTest {
-
+    /**
+     * Test to see if UserRegister successfully creates a User into Database
+     */
     @Test
     public void testCreateUserSuccess() throws IOException, ParseException {
         DataBaseController dataBaseController = new DataBaseController();
@@ -24,6 +26,9 @@ class UserRegisterTest {
         dataBaseController.deleteUser("StarlightUser");
     }
 
+    /**
+     * Test to see if the new User created with UserRegister UseCase is able to log in
+     */
     @Test
     public void testLoginNewUserSuccess() throws IOException, ParseException {
         DataBaseController dataBaseController = new DataBaseController();
@@ -36,15 +41,17 @@ class UserRegisterTest {
         dataBaseController.deleteUser("StarlightUser");
     }
 
+    /**
+     * Test to see if the user attempts to create a User with an existing user's username, the system returns the
+     * appropriate message
+     */
     @Test
     public void testUserExistsFail(){
         UserRegisterInputs inputs = new UserRegisterInputs("Herman1","Fuzzy321");
         UserRegisterResponseFormatter formatter = new UserRegisterResponseFormatter();
         UserRegister register = new UserRegister(formatter);
 
-        Exception exception = assertThrows(RuntimeException.class, () -> {
-            register.create(inputs);
-        });
+        Exception exception = assertThrows(RuntimeException.class, () -> register.create(inputs));
 
         String expectedMessage = "User already exists, pick a new username.";
         String actualMessage = exception.getMessage();
@@ -52,15 +59,17 @@ class UserRegisterTest {
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
     }
 
+    /**
+     * Test to see if the user attempts to create a User with a username over 15 characters, the system returns the
+     * appropriate message
+     */
     @Test
     public void testUsernameTooLongFail(){
         UserRegisterInputs inputs = new UserRegisterInputs("Herman10000000000000000000","Fuzzy321");
         UserRegisterResponseFormatter formatter = new UserRegisterResponseFormatter();
         UserRegister register = new UserRegister(formatter);
 
-        Exception exception = assertThrows(RuntimeException.class, () -> {
-            register.create(inputs);
-        });
+        Exception exception = assertThrows(RuntimeException.class, () -> register.create(inputs));
 
         String expectedMessage = "Username is too long. Username must have at most 15 characters.";
         String actualMessage = exception.getMessage();
@@ -68,15 +77,17 @@ class UserRegisterTest {
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
     }
 
+    /**
+     * Test to see if the user attempts to create a User with a username that is under 3 characters, the system returns
+     * the appropriate message
+     */
     @Test
     public void testUsernameTooShortFail(){
         UserRegisterInputs inputs = new UserRegisterInputs("A","Fuzzy321");
         UserRegisterResponseFormatter formatter = new UserRegisterResponseFormatter();
         UserRegister register = new UserRegister(formatter);
 
-        Exception exception = assertThrows(RuntimeException.class, () -> {
-            register.create(inputs);
-        });
+        Exception exception = assertThrows(RuntimeException.class, () -> register.create(inputs));
 
         String expectedMessage = "Username is too short. Username must have at least 3 characters.";
         String actualMessage = exception.getMessage();
@@ -84,15 +95,17 @@ class UserRegisterTest {
         Assertions.assertTrue(actualMessage.contains(expectedMessage));
     }
 
+    /**
+     * Test to see if the user attempts to create a password that is under 3 characters, the system returns the
+     * appropriate message
+     */
     @Test
     public void testPasswordTooShortFail(){
         UserRegisterInputs inputs = new UserRegisterInputs("StarlightUser1","1");
         UserRegisterResponseFormatter formatter = new UserRegisterResponseFormatter();
         UserRegister register = new UserRegister(formatter);
 
-        Exception exception = assertThrows(RuntimeException.class, () -> {
-            register.create(inputs);
-        });
+        Exception exception = assertThrows(RuntimeException.class, () -> register.create(inputs));
 
         String expectedMessage = "Password is too short. Password must have at least 3 characters.";
         String actualMessage = exception.getMessage();
